Adding full bruteforce feature
Hi,

I wanted to brute force subdomains with all possibilities and not just from a word list. With the added force_enum() function I set the default bruteforce length to three (3) for an optimal delay, but it can be use to brute force subdomain on of any length if time permits.

The function take the length (-l --length) and the charset (-k, --charset) 0 for alphanum, 1 for alpha, 2 for numeric, being alphanumeric by default.

This is rather useful for red teaming or bug bounty hunting to find hidden or forgotten subdomains that might host vulnerable code or services.  

Let me know if you have any question and if that make sense

Cheers