Update Python versions in DNSrecon workflow
This commit updates the Python versions used in the DNSrecon GitHub Actions workflow. Python 3.7 has been removed from the version matrix, and 3.11 has been added, to reflect the most current versions in use. This is important to ensure compatibility and optimal functionality of the DNSrecon tool on all supported Python platforms.