Add custom flake8 configuration
A custom .flake8 config file was added to ignore certain PEP 8 violations that are not applicable to this project. These are E203, E501, F405, F403, F401, E402, and W503. This configuration is now used when running flake8 in the GitHub workflow. This change makes linting more customized to the project's style guidelines.