Fix some problems within brute_tlds function
The latest version on master tests only all combination of {basename}.{cc}.{tld}, but not the single {basename}.{cc} or {basename}.{tld}.
Furthermore, most of the {cc}.{tld} are not valid combos, so there is a big wast of time. 
I suggest to download the list taken from https://publicsuffix.org/, which is a complete and maintained list of all the possibile tlds.